refactor(control-panel): split EnterpriseAdminPage into modular components
The CI file-size gate failed because EnterpriseAdminPage.tsx grew to
2763 lines, past the hard 2000-line limit and past the legacy-file
growth tolerance. Extract it into src/pages/enterprise-admin/:

- shared.tsx: constants, small helpers, GovernanceStatusStrip, Field
- forms.tsx: the six creation-drawer forms
- panels.tsx: the ten read/monitoring panels

EnterpriseAdminPage.tsx now only wires state, hooks, and layout,
importing the pieces above. No behavior change; verified with
typecheck, build, the full test suite (87/87), the file-size gate
script, and a live Playwright pass confirming the org-unit creation
drawer still opens.
